银行输入框不可点修复
