+++ wp/Other/Info.plist
@@ -44,7 +44,7 @@
 		</dict>
 	</array>
 	<key>CFBundleVersion</key>
-	<string>20170512</string>
+	<string>20170518</string>
 	<key>Fabric</key>
 	<dict>
 		<key>APIKey</key>
