--- wp/Scenes/Share/RechageVC/BankCardVC.swift
@@ -1,178 +1,0 @@
-//
-//  BankCardVC.swift
-//  wp
-//
-//  Created by sum on 2017/1/11.
-//  Copyright © 2017年 com.yundian. All rights reserved.
-//
-
-import UIKit
-import SVProgressHUD
-class BindingBankVCCell: UITableViewCell {
-    // 银行名称
-    @IBOutlet weak var bankName: UILabel!
-    // 银行名称
-    @IBOutlet weak var cardNum: UILabel!
-    
-    @IBOutlet weak var banklogo: UIImageView!
-    //银行背景
-    @IBOutlet weak var bankBg: UIImageView!
-    // 刷新cell
-    func update(_ data: Any!) {
-        
-        if let model = data as? BankListModel{
-            bankBg.backgroundColor =    BankLogoColor.share().readfilefromlocal(string: model.bank)
-            bankName.text = model.bank
-            let index = model.cardNo.index(model.cardNo.startIndex,  offsetBy: 4)
-            let index1 = model.cardNo.index(model.cardNo.startIndex,  offsetBy: model.cardNo.length()-3)
-            cardNum.text =  model.cardNo.substring(to: index) + "  ****   ****   *** " + model.cardNo.substring(from: index1)
-            banklogo.image = BankLogoColor.share().checkLocalBank(string: model.bank) ? UIImage.init(named: BankLogoColor.share().checkLocalBankImg(string: model.bank)) : UIImage.init(named: "unionPay")
-        }
-    }
-}
-class BankCardVC: BaseListTableViewController {
-    
-    var dataArry : [BankListModel] = []
-    override func viewDidLoad() {
-        
-        self.title = "我的银行卡"
-        super.viewDidLoad()
-        
-        
-    }
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        
-        self.didRequest()
-    }
-    override func didRequest() {
-        AppAPIHelper.user().bankcardList(complete: { [weak self](result) -> ()? in
-            
-            if let object = result {
-                let Model : BankModel = object as! BankModel
-                self?.didRequestComplete(Model.cardList as AnyObject)
-                self?.dataArry = Model.cardList!
-                self?.tableView.reloadData()
-            }else {
-                self?.didRequestComplete(nil)
-            }
-            return nil
-            }, error: errorBlockFunc())
-        
-        
-    }
-    //MARK: 实现银行卡左滑删除的代理
-    override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
-        
-        
-        let share = UITableViewRowAction(style: .normal, title: "删除") { action, index  in
-            
-
-            self.UnbindCard(number: indexPath.section)
-            ShareModel.share().shareData["number"] = "\(indexPath.section)"
-            self.dataArry.remove(at: Int(ShareModel.share().shareData["number"]!)!)
-            tableView.reloadData()
-
-//            ShareModel.share().shareData["number"] = "\(indexPath.section)"
-//            
-//             self.UnbindCard(number: indexPath.section)
-
-        }
-        
-        share.backgroundColor = UIColor.red
-        
-        return [share]
-    }
-    
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        
-        return dataArry.count
-    }
-    override func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat{
-        
-        return 15
-    }
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int{
-        
-        return 1
-    }
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        return true
-    }
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell
-    {
-        let cell : BindingBankVCCell = tableView.dequeueReusableCell(withIdentifier: "BankCardVCCell") as! BindingBankVCCell
-        
-        let  Model : BankListModel = self.dataArry[indexPath.section]
-        
-        cell.update(Model.self)
-        return cell
-        
-    }
-   
-    //MARK: 解绑逻辑
-    func UnbindCard ( number: Int) {
-        
-    }
-    deinit {
-        ShareModel.share().shareData.removeAll()
-    }
-    //输入密码 的代理方法
-    override func alertView(_ alertView: UIAlertView, clickedButtonAt buttonIndex: Int) {
-        
-        
-        if buttonIndex==0{
-            
-            if buttonIndex==0{
-                
-//                let  model :BankListModel = self.dataArry[Int(ShareModel.share().shareData["number"]!)!] as BankListModel
-                
-//                AppAPIHelper.user().unbindcard(vToken: UserModel.share().codeToken, bid: Int32(model.bid), timestamp:  Int64(UserModel.share().timestamp) , phone: (UserModel.share().getCurrentUser()?.phone)!, vCode: (alertView.textField(at: 0)?.text)!, complete: { [weak self](result) -> ()? in
-//                    
-//                
-//    //
-//                    //                                                self?.tableView.reloadData()
-//                                        if result != nil{
-//                                            
-//                                            let dic : NSDictionary = result as! NSDictionary
-//                                            if (dic.object(forKey: "error") != nil){
-//                                                let error : Int = dic.object(forKey: "error") as! Int
-//                                                if error == -113{
-//                                                
-//                                                    SVProgressHUD.showError(withStatus: "验证码错误")
-//                                                     self?.tableView.reloadData()
-//                                                    return nil
-//                                                }
-//                                            }
-//                                            self?.didRequest()
-//                                            self?.dataArry.remove(at: Int(ShareModel.share().shareData["number"]!)!)
-//                                            self?.tableView.reloadData()
-//                                                                }else{
-//                    
-//                                                                }
-//                    
-//                    return nil
-//                    }, error: errorBlockFunc())
-//                //                AppAPIHelper.user().unbindcard(bankId: model.bid, vCode: (alertView.textField(at: 0)?.text)!, complete: { [weak self](result) -> ()? in
-//                //
-//                //                    if result != nil{
-//                //                        self?.dataArry.remove(at: Int(ShareModel.share().shareData["number"]!)!)
-//                //
-//                //                        self?.tableView.reloadData()
-//                //                    }else{
-//                //
-//                //                    }
-//                //
-//                //                    
-//                //                    //            self?.didRequest()
-//                //                    
-//                //                    return nil
-//                //                    }, error: errorBlockFunc())
-//            }
-        }
-        
-        }
-    }
-    
-}
